deploy: Google Chrome origin trial token
Due to a change in the policy of Google Chrome, is no longer possible
request an http local resource from an http page.

This change doesn't allow anymore to call the local chilli instance from
the authentication portal and the users cannot request the access to
internet even if the authentication was successful.

As a temporary workaround, a token can be requested to disabling this
feature in Google Chrome browsers.

See https://developer.chrome.com/blog/private-network-access-update/#register-deprecation-trial