dedalo. add new users auth method
Starting from version 94 Chrome block any requests to private networks
from insecure public websites[1] and this new behavior interfere with
the captive portal authentication flow.

An authenticated user at the end of the authentication process, request
access to the internet by making a call to the local `coova-chilli`
instance, and this communication is no longer possible.

To resolve the above problem, in the new auth the local dedalo instance
will periodically ask the status of the connected devices and allowing
internet access to the authenticated ones.

[1]https://developer.chrome.com/blog/private-network-access-update